Extending Figures trait and its abilities

+++ src/main/scala/Areable.scala
@@ -1,5 +1,5 @@
 trait Areable {
 
-  def area: Long
+  def area: Double
 
 }
